chore: rename prisma module to db
